python/maf04_aifoundry_v1/pyproject.toml
[project]
name = "maf04-aifoundry-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-azure-ai>=1.0.0b251120",
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.1",
]
